+++ src/activation/sigmoid.cpp
@@ -4,11 +4,11 @@
 
 Sigmoid::Sigmoid() {}
 
-float Sigmoid::function(float x) {
+double Sigmoid::function(double x) {
     return 1.0f / (1.0f + std::exp(-x));
 }
 
-float Sigmoid::derivative(float x) {
-    float fx = function(x);
+double Sigmoid::derivative(double x) {
+    double fx = function(x);
     return fx * (1.0f - fx);
 }
